Expanded underrepresented group examples for scholarship
This expands the list of examples of underrepresented groups to include many more examples. This list is mostly adopted from AlterConf with a few tweaks. I'm not fixated on every word of this, we may need to make more tweaks.

The idea behind this is that I think there are people from underrepresented groups, especially those not visible or not often mentioned, that would not feel welcome in the scholarship program because they are not included in our list. We do already have a "not limited to" clause, but listing groups explicitly is a much stronger message that those people are welcome to apply.